Remove useless file
